feat: 支持GitHub Re-run #3676

+++ src/backend/ci/ext/tencent/external/biz-external/src/main/kotlin/com/tencent/devops/external/resources/ServiceGithubResourceImpl.kt
@@ -76,7 +76,7 @@ class ServiceGithubResourceImpl @Autowired constructor(
     override fun updateCheckRuns(
         accessToken: String,
         projectName: String,
-        checkRunId: Int,
+        checkRunId: Long,
         checkRuns: GithubCheckRuns
     ): Result<Boolean> {
         githubService.updateCheckRuns(accessToken, projectName, checkRunId, checkRuns)
